+++ SogeVet.Server/Models/OrderItemDto.cs
@@ -6,7 +6,7 @@ namespace SogeVet.Server.Models
     {
         public int Id { get; set; }
         public int Quantity { get; set; }
-        public int UnitPrice { get; set; }
+        public float UnitPrice { get; set; }
 
         public int OrderId { get; set; }
